move New*Message apis to TransportFactory
  * add facades to Device and Channel
  * add Transport() getter to Device
  * inline Transport() getter from Channel
