+++ fairmq/FairMQTransportFactory.h
@@ -35,11 +35,19 @@ class FairMQTransportFactory
     /// Initialize transport
     virtual void Initialize(const FairMQProgOptions* config) = 0;
 
-    /// Create an empty message (e.g. for receiving)
+    /// @brief Create empty FairMQMessage
+    /// @return pointer to FairMQMessage
     virtual FairMQMessagePtr CreateMessage() const = 0;
-    /// Create an empty of a specified size
+    /// @brief Create new FairMQMessage of specified size
+    /// @param size message size
+    /// @return pointer to FairMQMessage
     virtual FairMQMessagePtr CreateMessage(const size_t size) const = 0;
-    /// Create a message from a supplied buffer, size and a deallocation callback
+    /// @brief Create new FairMQMessage with user provided buffer and size
+    /// @param data pointer to user provided buffer
+    /// @param size size of the user provided buffer
+    /// @param ffn callback, called when the message is transfered (and can be deleted)
+    /// @param obj optional helper pointer that can be used in the callback
+    /// @return pointer to FairMQMessage
     virtual FairMQMessagePtr CreateMessage(void* data, const size_t size, fairmq_free_fn* ffn, void* hint = nullptr) const = 0;
 
     /// Create a socket
@@ -63,6 +71,50 @@ class FairMQTransportFactory
     virtual ~FairMQTransportFactory() {};
 
     static auto CreateTransportFactory(const std::string& type) -> std::shared_ptr<FairMQTransportFactory>;
+
+    static void FairMQNoCleanup(void* /*data*/, void* /*obj*/)
+    {
+    }
+
+    template<typename T>
+    static void FairMQSimpleMsgCleanup(void* /*data*/, void* obj)
+    {
+        delete static_cast<T*>(obj);
+    }
+
+    template<typename T>
+    inline FairMQMessagePtr NewSimpleMessage(const T& data) const
+    {
+        // todo: is_trivially_copyable not available on gcc < 5, workaround?
+        // static_assert(std::is_trivially_copyable<T>::value, "The argument type for NewSimpleMessage has to be trivially copyable!");
+        T* dataCopy = new T(data);
+        return CreateMessage(dataCopy, sizeof(T), FairMQSimpleMsgCleanup<T>, dataCopy);
+    }
+
+    template<std::size_t N>
+    inline FairMQMessagePtr NewSimpleMessage(const char(&data)[N]) const
+    {
+        std::string* msgStr = new std::string(data);
+        return CreateMessage(const_cast<char*>(msgStr->c_str()), msgStr->length(), FairMQSimpleMsgCleanup<std::string>, msgStr);
+    }
+
+    inline FairMQMessagePtr NewSimpleMessage(const std::string& str) const
+    {
+
+        std::string* msgStr = new std::string(str);
+        return CreateMessage(const_cast<char*>(msgStr->c_str()), msgStr->length(), FairMQSimpleMsgCleanup<std::string>, msgStr);
+    }
+
+    template<typename T>
+    inline FairMQMessagePtr NewStaticMessage(const T& data) const
+    {
+        return CreateMessage(data, sizeof(T), FairMQNoCleanup, nullptr);
+    }
+
+    inline FairMQMessagePtr NewStaticMessage(const std::string& str) const
+    {
+        return CreateMessage(const_cast<char*>(str.c_str()), str.length(), FairMQNoCleanup, nullptr);
+    }
 };
 
 #endif /* FAIRMQTRANSPORTFACTORY_H_ */
